--- .gitignore
@@ -1,19 +1,0 @@
-# Media Files
-# *.png
-*.jpg
-*.bmp
-*.tif
-*.avi
-*.mp4
-
-# Documnets
-*.docx
-*.doc
-*.xlsx
-*.xls
-*.pptx
-*.ppt
-# *.pdf
-
-# MATLAB Files
-*.asv
