+++ Inputs/logspace.m
@@ -1,0 +1,2 @@
+round(logspace(0.5,3,20)) % full
+round(logspace(0.5,3,10)) % partial
