+++ JuliaChangelog.txt
@@ -3,7 +3,7 @@ New Changelog:
 Release Notes:
 
 	2.0: 
-		* Combined separated function files JuliaMatrixBenchmark and JuliaMatrixBenchmarkOp into JuliaBench and JuliaBenchOp respectively for easier access
+		* Combined separated function files JuliaMatrixBenchmark and JuliaMatrixBenchmarkOp into JuliaBench and JuliaBenchSIMD respectively for easier access
 		* Used @benchmark instead of deprecated and inaccurate tic() and toq(), so all functions now just include the algorithm or operation. Benchmark result may not be comparable to ones in previous versions.
 		* @benchmarks calls allFunctions in local scope (using $). Then the minimum run time among different benchmarks for each function is chosen.
 		* Removed totalRunTime because it does not mean anything anymore
@@ -13,12 +13,17 @@ Release Notes:
 		* Removed rounds because readdlim supports converting to int64 itself
 		* Changed deprecated squeeze to droplims
 		* jj for loop changed to be like fun in allFunctions. This allows simply adjusting allFunctions array if want to remove some functions from benchmark
-		* Added dims= to 2nd input argument of minimum(), maximum(), and sum() functions
+		* Added dims= to 2nd input argument of minimum(), maximum(), findmin() and sum() functions
 		* Added dummy mY to functions that don't use it. (for consistency in code logic)
 		* Changed deprecated expm with exp
 		* Changed deprecated sqrtm with sqrt
-		* eig changed with eigvals. May need another function for eigvectors
-
+		* eig changed with eigen. 
+		* vClusterId modified to be compatible with Cartesian indexing
+		* mRunTime is stored in a mat file using MAT package.
+		* tRunTime is a table which has the run time information in a intuitive manner
+		* added 0 benchmark operation mode which is for fast testing only
+		* changed file and folder management
+		* now average of different iteration of Run Times is calculated (for different kk)
 
 Old changelog:
