+++ MatlabMain.m
@@ -1,6 +1,8 @@
-function tRunTime = MatlabMain( operationMode )
+clc;clear;
 
-tRunTime = MatlabBench(operationMode);
+operationMode=2;% 0 for test only # 1 for partial benchmark # 2 for full benchmark
 
-end
+[ mRunTime,tRunTime ] = MatlabBench(operationMode);
 
+writetable(tRunTime,fullfile('RunTimeData', 'RunTimeMatlabTable.csv'),'WriteVariableNames',false);
+save(fullfile('RunTimeData', 'RunTimeMatlab.mat'),'mRunTime')
