--- RunTimeData/RunTimeJulia0001.csv
@@ -1,14 +1,0 @@
-904e-9,1808e-9,904e-9,1205e-9,3013e-9,5122e-9
-904e-9,1205e-9,904e-9,904e-9,2712e-9,5122e-9
-1808e-9,241e-8,2109e-9,1205e-9,3013e-9,1175e-8
-5422e-9,3012e-9,3314e-9,1205e-9,3917e-9,22595e-9
-24402e-9,20486e-9,3344e-8,1808e-9,1416e-8,167503e-9
-122915e-9,74412e-9,291322e-9,143101e-9,46395e-9,618796e-9
-460029e-9,281079e-9,926387e-9,107551e-9,175035e-9,0.002511033
-0.001060449,746531e-9,0.001798846,115384e-9,40279e-8,0.005875548
-0.003002696,0.0020739,0.005273924,145209e-9,0.001098107,0.01958998
-0.006564539,0.007453266,0.013413168,171721e-9,0.002582735,0.047575278
-0.01403558,0.011059093,0.024377965,203955e-9,0.004540347,0.147672888
-0.116302288,0.10627412,0.148354045,0.002251646,0.021685571,0.621290313
-0.182242306,0.168405561,0.453054062,0.005154925,0.048494736,1.110252911
-0.274058776,0.245145585,0.906215976,0.008219079,0.085424254,1.589702201
