--- RunTimeData/RunTimeJulia0002.csv
@@ -1,14 +1,0 @@
-50913e-9,15364e-9,4218e-9,32235e-9,1205e-9,9038e-9
-127435e-9,26813e-9,12352e-9,15364e-9,1205e-9,22294e-9
-105443e-9,97609e-9,42177e-9,36754e-9,2109e-9,5965e-8
-284695e-9,354286e-9,159067e-9,124422e-9,3314e-9,360915e-9
-0.001678642,0.001915435,0.002010333,0.00102972,106346e-9,0.002624007
-0.009160529,0.004881377,0.004496965,0.009122871,209378e-9,0.007083013
-0.013224577,0.01431937,0.020709777,0.078578939,618193e-9,0.02389263
-0.030462891,0.024408694,0.04148975,0.154022625,0.001293928,0.047931674
-0.155371986,0.068527272,0.124056818,0.308065435,0.002355882,0.140241613
-0.293248377,0.174482052,0.303503999,0.83926341,0.004985012,0.329712142
-0.741371352,0.338720232,0.636193231,1.388697788,0.011945411,0.628738457
-3.393030563,1.828654406,3.512926024,6.782945758,0.053805715,3.868881859
-8.251872204,5.273375209,10.917288324,21.452609999,0.138239414,12.095518701
-17.001981412,11.650788257,24.214127766,42.750310151,0.251438384,26.736623435
