--- RunTimeData/RunTimeJulia0003.csv
@@ -1,14 +1,0 @@
-11448e-9,15063e-9,3917e-9,9309e-8
-2621e-8,30729e-9,4519e-9,94597e-9
-45189e-9,43984e-9,6025e-9,102129e-9
-219621e-9,22173e-8,12955e-9,113878e-9
-666697e-9,0.001410216,80438e-9,151536e-9
-0.004424058,0.013153479,287406e-9,294636e-9
-0.052753702,0.047848524,0.001419254,877582e-9
-0.082943949,0.087783452,0.002923464,0.001971771
-0.159001913,0.177412144,0.00944703,0.004985012
-0.310443913,0.43852441,0.025368521,0.012141835
-0.493967193,0.568671577,0.044788889,0.02321328
-0.694115414,0.8207628,0.384742495,0.12459337
-1.337682076,2.143543465,0.837971892,0.419156461
-2.014445298,3.795277087,1.500692003,0.788423939
