--- RunTimeData/RunTimeJuliaOpt0001.csv
@@ -1,14 +1,0 @@
-904e-9,904e-9,1206e-9,1205e-9,3314e-9,603e-9
-903e-9,602e-9,1808e-9,1205e-9,3013e-9,1205e-9
-3314e-9,904e-9,2711e-9,2109e-9,5121e-9,5724e-9
-8435e-9,1205e-9,5724e-9,1506e-9,4218e-9,9942e-9
-3344e-8,12653e-9,58144e-9,6025e-9,23197e-9,110865e-9
-225345e-9,35851e-9,20275e-8,167503e-9,78328e-9,487445e-9
-707065e-9,125928e-9,688688e-9,180456e-9,189194e-9,0.001044481
-0.001012848,206667e-9,0.00174793,103334e-9,376278e-9,0.002333285
-0.002876162,649825e-9,0.005117864,167201e-9,0.00130658,0.006584113
-0.006520546,0.001646103,0.013330606,227755e-9,0.00257038,0.014852589
-0.013433939,0.003230749,0.02537783,370855e-9,0.004484005,0.026995313
-0.051701626,0.016145612,0.156419894,0.002411011,0.020410297,0.180672805
-0.181419939,0.034227443,0.462186618,0.005257047,0.048005124,0.325634747
-0.2713769,0.061120929,0.944704602,0.008615833,0.08292728,0.512411752
