--- RunTimeData/RunTimeJuliaOpt0002.csv
@@ -1,14 +1,0 @@
-45792e-9,15967e-9,3615e-9,6929e-9,1205e-9,13557e-9
-87668e-9,28017e-9,12653e-9,17473e-9,1205e-9,19582e-9
-94898e-9,91283e-9,5001e-8,34043e-9,2108e-9,56638e-9
-255472e-9,339223e-9,149427e-9,121409e-9,3615e-9,300661e-9
-0.001002907,0.002029915,0.001820537,971878e-9,78028e-9,0.002271228
-0.004404476,0.006107822,0.005950261,0.009842891,245831e-9,0.009806137
-0.016302891,0.015140615,0.02491542,0.101872054,682062e-9,0.037112688
-0.025721602,0.024863603,0.047331556,0.155612996,0.001478302,0.058865441
-0.082253453,0.094531761,0.182459518,0.35488213,0.003695001,0.149434679
-0.355886845,0.171821892,0.341788603,0.833156792,0.005217287,0.41028364
-0.531481766,0.317455225,0.569682619,1.444515945,0.008484793,0.607150798
-3.23328965,1.653344781,3.298342748,6.512160212,0.040600418,3.616668022
-7.643903777,4.987558709,10.264815099,21.330646662,0.126866706,11.409008088
-15.583676021,10.932401524,22.818086673,42.723749228,0.234987272,25.117706178
