--- RunTimeData/RunTimeJuliaOpt0003.csv
@@ -1,14 +1,0 @@
-15365e-9,39466e-9,4519e-9,83149e-9
-30127e-9,26812e-9,4218e-9,103936e-9
-62964e-9,68086e-9,8134e-9,97007e-9
-254266e-9,283188e-9,19582e-9,182265e-9
-932713e-9,0.001052917,126531e-9,217813e-9
-0.003266302,0.008057601,33139e-8,335005e-9
-0.019055237,0.029500656,0.001231566,871858e-9
-0.032209921,0.045960505,0.002988236,0.002006718
-0.072876617,0.056408634,0.010158314,0.005038637
-0.059430009,0.088249808,0.02491542,0.011751698
-0.102461627,0.143262988,0.046855258,0.022822541
-0.415653968,0.928580295,0.385454682,0.119280883
-0.828750207,1.625202524,0.819282691,0.404206245
-1.626159941,3.321303568,1.424654827,0.75537771
