+++ RunTimeData/RunTimeJuliamklSIMDTable.csv
@@ -1,0 +1,4 @@
+FunctionName\MatrixSize,2,4,6,8,11,14,20,25,35,50,65,85,105,120,130,150,180,220,300,400,550,740,1000
+Matrix Addition,0.04642664353995557,0.04997331560337687,0.06092596760842698,0.09233640146901037,0.1663961405687296,0.2798114046035904,1.0608972348231205,0.84066806447265,1.5007285714285714,8.414747222222223,7.806491071428572,9.19828125,19.6,23.575,20.48725,27.3005,38.025000000000006,51.45,107.7625,624.850375,1307.1,2455.100375,5061.162625000001
+Matrix Multiplication,0.045301256281407035,0.05364150705645161,0.05857618771164955,0.09066201716126235,0.16906192640897172,0.27916717239348876,0.5733791893320629,0.8758461867484336,1.5225618095930233,8.531905555555557,7.939,9.752062500000001,15.641875,18.725,21.287375,25.175375,35.525125,55.26275,102.187625,635.4627500000001,1316.199625,2531.36225,5119.175125
+Element Wise Operations,0.3979681932102985,0.8588506163001022,2.5312,4.06875,7.037484375,11.32075,10.725000000000001,16.374750000000002,30.849875,112.025125,125.73775,200.66262500000002,286.925,366.28749999999997,434.499875,573.750125,855.474875,1372.46225,2586.325125,8799.650375000001,18178.1125,32344.849874999996,59688.28725
