+++ RunTimeData/RunTimeJuliamklTable.csv
@@ -1,0 +1,17 @@
+FunctionName\MatrixSize,2,4,6,8,11,14,20,25,35,50,65,85,105,120,130,150,180,220,300,400,550,740,1000
+Matrix Generation,0.11734775250635351,0.22874603276316335,0.5297003847972876,0.8741100765498314,2.147475,3.4175125,3.5843749999999996,6.3001249999999995,9.92525,29.050625000000004,43.199999999999996,72.87525,110.099875,143.35,168.462625,258.75025000000005,315.975375,484.450125,740.861875,2152.475,4132.337625,7836.187499999999,18012.13775
+Matrix Addition,0.05486098484848485,0.06685788972063787,0.09509368296341245,0.12072298785211472,0.17699112272976558,0.274692467549858,0.47066331334587547,0.724320092900739,1.3483148121317492,3.025013888888889,5.719933035714286,10.138333333333334,14.249875,18.9745,23.7745,28.374625,38.375,55.212624999999996,114.600125,568.962375,1199.2625,2434.65025,5846.7875
+Matrix Multiplication,0.12266764269526292,0.5709883244834315,0.39394890339648403,0.4991155761162333,0.6935141003126882,0.9431970916125647,1.8475125,3.094486111111111,7.09925,13.925125,23.662625,44.313,66.7,84.42499999999998,115.80025,143.987875,239.17487500000001,442.037375,1122.2997500000001,3877.1996249999997,8972.537499999999,21726.512499999997,71316.349625
+Matrix Quadratic Form,0.5024820955017473,0.4979488259588777,0.6699908888190955,0.6937826633165829,0.7528814599914535,0.8017039559925094,0.715639317745229,0.9404398210658728,1.8825249999999998,2.565025,3.741722222222222,5.240625,10.945,9.772575,10.78455,11.295770833333334,14.925,20.887,25.024874999999998,34.724875,66.27475000000001,160.82525,1029.4873750000002
+Matrix Reductions,2.622236111111111,2.8610833333333336,3.063861111111111,3.305541666666667,3.7847222222222223,4.415625,4.043749999999999,4.5107678571428576,6.0624166666666675,9.126083333333334,17.312124999999998,28.862375,29.424999999999997,36.0255,42.225249999999996,53.99975,73.3755,106.438,187.75025,764.649875,1532.9123749999999,2895.7376249999998,7623.2119999999995
+Element Wise Operations,1.1075374999999998,2.6625,4.262472222222223,6.403109375,10.65825,16.43115625,14.37575,21.15,40.075,83.350375,146.93725,326.17487500000004,373.888,486.812625,586.85025,781.51275,1073.4250000000002,1707.8,3410.3872499999998,14861.94975,25284.262625000003,45402.57525,105503.175
+Matrix Exponential,3.538902777777778,4.88128125,7.098196428571429,10.579208333333334,18.625,26.675,24.3625,42.2375,92.1875,202.35025000000002,354.024625,611.36275,1025.487125,1210.112375,1969.6,2730.6125,4162.7375,7731.77525,18006.92525,54995.0,128402.14924999999,271750.3995,805312.9751250001
+Matrix Square Root,1.544975,7.249958333333334,11.441708333333334,15.325000000000001,24.237375,32.95,49.150749999999995,68.900375,134.18737499999997,283.70012499999996,481.825,794.662375,1233.0998749999999,1683.51225,2000.45,2684.4996250000004,3799.58775,5705.63775,11510.7,26143.062375,54439.475125,110406.73712499999,269044.66225
+Singular Value Decomposition,2.039975,5.75,11.041708333333334,15.725000000000001,21.175,32.20025,58.16250000000001,87.70025,182.21250000000003,503.525125,759.999875,1318.487375,1757.07475,2167.4873749999997,2658.7373749999997,4034.12525,5303.625,8611.862625,19579.82475,34895.2375,73206.14924999999,158768.04962499999,449854.52425
+Eigen Decomposition,2.6861527777777776,6.479104166666667,11.8495,18.475125,35.300125,52.900125,95.0745,161.325125,440.61287500000003,1014.2375,2194.587625,4812.7125,7535.900125,9463.861875,11829.737625,18855.899875000003,25925.51225,38139.949875,65302.7125,124001.54974999999,317167.375,578886.1126250001,1.0931140995e6
+Cholesky Decomposition,0.21299913995726494,0.891544289294408,1.2644511092083137,1.5587531512605044,2.4274750000000003,3.54995,3.0889027777777778,4.721482142857143,10.950125,21.987499999999997,39.962375,65.599875,97.72512499999999,135.94975,151.787125,198.32525,269.650125,441.20025000000004,970.7252500000001,3084.287625,7210.662625,15659.650374999997,37272.10025
+Matrix Inversion,4.32965625,10.7125,16.549625,23.1125,31.15,44.675125,74.86250000000001,114.2505,401.76275000000004,638.13775,955.85,1949.862375,2880.2252500000004,3463.137625,4349.21225,5618.225,8186.124625,13131.212375,25429.62525,55265.200000000004,114439.91287500001,238300.362875,521033.07525
+Linear System Solution,0.6429440212926993,1.109658931921068,2.315,4.136069444444445,7.791035714285714,10.670100000000001,17.099875,25.974875,72.324625,112.425125,152.387625,266.150125,430.36300000000006,524.9000000000001,855.9124999999999,1019.9375,1461.850125,2375.562375,4527.55,12191.987625,28906.499875,56053.86275,116484.79999999999
+Linear Least Squares,0.8768108152180853,3.266652777777778,4.727777777777778,6.989249999999999,12.082506250000002,16.308208333333333,21.850499999999997,32.58725,90.00037499999999,124.162875,228.7875,317.16225,562.92475,755.2626250000001,1073.77525,1365.062625,2006.0375,3260.79975,6820.687249999999,18997.862500000003,45193.624875,91175.37512499999,336527.51250000007
+Squared Distance Matrix,1.0366056961534869,2.039975,2.6399875,3.9087624999999995,6.106944444444445,3.937546875,7.657525000000001,10.499749999999999,27.437124999999998,48.450125,80.262875,128.68737499999997,228.35,305.225,357.80037500000003,495.087625,757.46275,1348.225,2922.9248750000006,9010.824499999999,20979.062124999997,39132.062375,108354.41275
+K-Means,22.4005,24.225375,26.612375,30.950375,33.837,30.14975,31.387999999999998,35.625,40.949875,46.524875,56.40025,70.112125,104.237,123.68787499999999,121.96300000000001,255.91224999999997,314.487875,423.97475000000003,841.662375,1955.4626250000001,4889.762625,10040.662250000001,24898.875625
