--- RunTimeData/RunTimeMatlab0001.csv
@@ -1,14 +1,0 @@
-4.2779e-05,6.0253e-07,3.4947e-05,3.9164e-06,2.7114e-05,3.2235e-05
-1.0845e-05,6.0253e-07,6.0253e-06,4.2177e-06,9.3392e-06,8.4354e-06
-5.1215e-06,9.0379e-07,2.1088e-06,4.2177e-06,3.3139e-06,6.3265e-06
-9.9417e-06,9.0379e-07,3.6152e-06,3.9164e-06,3.9164e-06,2.3499e-05
-4.6093e-05,3.0126e-06,1.5967e-05,5.4227e-06,9.9417e-06,7.2303e-05
-0.00019673,1.1147e-05,4.6997e-05,1.6268e-05,5.0612e-05,0.0001681
-0.0011816,2.5005e-05,0.0001675,3.0126e-05,7.2303e-05,0.00051727
-0.0027023,5.3926e-05,0.00056156,4.3081e-05,0.00011358,0.00098935
-0.005292,0.00075045,0.0036393,0.00010845,0.00058415,0.0030039
-0.01177,0.0011807,0.0095278,0.00013406,0.0013445,0.0066489
-0.020201,0.0018757,0.020277,0.00019221,0.0024881,0.012392
-0.083167,0.0096871,0.14158,0.0022519,0.008035,0.070849
-0.18711,0.020843,0.27154,0.0050028,0.019026,0.14647
-0.32944,0.036298,0.52847,0.0090936,0.036116,0.2439
