--- RunTimeData/RunTimeMatlab0002.csv
@@ -1,14 +1,0 @@
-0.00035188,0.00012171,4.5491e-05,4.5792e-05,1.5666e-05,0.00013708
-0.00016238,3.9767e-05,2.1088e-05,3.9164e-05,4.8202e-06,6.5374e-05
-0.0001907,4.4888e-05,4.2779e-05,4.7298e-05,1.8076e-06,9.4296e-05
-0.00021751,0.00011117,0.00018799,0.00026722,3.6152e-06,0.00018498
-0.00065344,0.0003344,0.00066549,0.0024547,1.1147e-05,0.00060976
-0.0014548,0.00086553,0.001988,0.006078,4.1273e-05,0.0018229
-0.005358,0.0027078,0.0055631,0.030216,0.00012864,0.0064546
-0.010338,0.0062009,0.012689,0.059956,0.00024674,0.015097
-0.042217,0.016122,0.033464,0.1434,0.00090711,0.044504
-0.11947,0.03989,0.080687,0.39669,0.0018937,0.10708
-0.25041,0.08239,0.16125,0.67349,0.0037396,0.20301
-1.7304,0.50545,1.4144,3.9207,0.018757,1.6865
-4.7521,1.7791,5.3008,12.588,0.054078,6.0536
-10.109,4.1806,10.27,27.185,0.12223,14.702
