--- RunTimeData/RunTimeMatlab0003.csv
@@ -1,14 +1,0 @@
-3.826e-05,7.3207e-05,4.3081e-05,8.104e-05
-4.0671e-05,3.2838e-05,1.2051e-05,8.345e-05
-1.5967e-05,3.103e-05,6.3265e-06,8.1341e-05
-2.9223e-05,6.5675e-05,1.3557e-05,8.7668e-05
-0.00012412,0.00016298,4.0671e-05,0.00014822
-0.000548,0.00052992,0.00016931,0.0003576
-0.0021185,0.0012822,0.00039827,0.0021212
-0.003294,0.0027087,0.00087728,0.0033838
-0.010062,0.0080446,0.0048377,0.0071122
-0.024304,0.032962,0.012536,0.016595
-0.034325,0.046801,0.0195,0.037532
-0.23168,0.31375,0.12483,0.15842
-0.64493,0.9305,0.34394,0.36786
-1.2929,2.044,0.72886,0.74474
