+++ RunTimeData/RunTimeMatlabTable.csv
@@ -1,0 +1,17 @@
+FunctionName\\MatrixSize,2,4,6,8,11,14,20,25,35,50,65,85,105,120,130,150,180,220,300,400,550,740,1000
+Matrix Generation,8.90776062247121,13.009387882754,9.09561372292756,9.73372073471233,10.6561577582667,11.7657912827462,15.0005561946734,18.9888859297565,29.3500020117138,50.3338,78.6511598290598,126.471975,187.313883333333,258.593366666667,281.275891666667,374.210683333333,528.21085,784.8316,1504.2861,3478.5361,6567.7861,12184.4611,22224.8611
+Matrix Addition,8.91858651647791,9.05149868111448,8.79912307225721,8.77902891066217,8.89022293661282,8.89673795547569,8.91679335130278,9.32214428904429,10.0227611243312,10.6945480175643,12.118565807373,13.9368243240775,18.2508874229701,20.3703307957393,21.1271257314466,21.6038598554422,24.2457535207505,24.6609291966685,38.3218302419355,457.223183333333,1070.83585,1893.6611,3547.4611
+Matrix Multiplication,10.073145848512,10.0164429206729,9.8373773712907,10.1622587027351,10.284342867375,10.4576544250768,10.879379804865,12.0874418847496,31.5101186491936,35.0088446949602,47.2382428571429,43.9455338768116,72.8757988095238,93.3551909090909,99.4098944444444,145.399480952381,207.711271428571,381.273183333333,882.67335,2740.8861,6686.8611,16134.6361,35359.9361
+Matrix Quadratic Form,10.4040781186574,9.92165652782511,10.0203150028535,10.0888664115964,10.1610589575065,10.2450952742715,10.4304552461597,10.9389531066156,11.3777423977252,11.9042249230237,13.1829413584475,14.6552824846356,22.5713413623414,23.0253860022577,23.8533032312925,21.8921183806147,38.1080627924445,26.0506329365079,51.1456491545093,54.9068236842105,99.1798291666667,276.532183333333,824.16085
+Matrix Reductions,9.78234682165225,9.74315540276271,9.66626454942123,9.77691098790322,9.98113019843707,10.0976297004466,10.3764347985348,10.9851375638788,12.5209959404749,15.0900989537177,18.8612664781297,24.5664821524966,37.735280651341,44.0991494117647,45.7212007905138,50.6606527777778,67.61353125,66.07665625,94.9317723484848,528.167683333333,1138.55685,2250.6861,3886.9611
+Element Wise Operations,10.986579479378,11.2084864418938,12.4331203615829,13.9688217566016,18.551526161103,21.5021599363197,41.5823221713472,47.8212561264822,93.5475000832501,125.390023611111,215.808616666667,347.924516666667,454.660766666667,517.6691,601.2441,806.97335,1090.3861,1588.4361,3195.5611,9614.3611,17362.7361,33249.7361,58760.4611
+Matrix Exponential,87.6517416666667,100.57095,92.9814045454546,128.1762875,135.343108928571,139.846983928571,151.731957142857,180.86555,575.6941,540.255266666667,739.78585,849.5026,1327.2361,1513.1861,2101.0861,3406.3861,4470.7611,7395.3111,14844.0111,33427.6861,70826.3611,204104.4111,491661.9611
+Matrix Square Root,23.0728990853659,28.6788878787879,32.0621447044335,33.9020808429119,59.561233492823,90.7717888888889,130.309475,128.277918181818,290.4606,385.6051,460.067683333333,628.3316,1102.23185,1490.3861,1667.9611,2080.6611,2700.5861,3853.9861,9771.5111,15409.3111,38835.4361,75793.9111,168695.6611
+SVD,13.7104965134222,17.3441048027444,28.3881878091547,26.2797895645646,46.933480952381,85.6962830065359,148.910097727273,147.724716666667,304.676216666667,394.798183333333,1002.15685,901.81935,1716.5361,2154.0111,2873.0611,2562.0361,3751.0361,5960.3361,15432.1611,21974.2611,70765.3111,112778.0861,319957.2111
+Eigen Decomposition,13.6725406325152,17.3896736853832,31.2524037206614,29.6803218963832,43.9386346889952,61.5995375,118.626473611111,167.832766666667,371.688433333333,772.0151,1420.5861,3234.5361,5203.8111,6547.0111,8484.9611,12695.7111,18520.4861,28286.2611,54528.2111,80253.1111,223028.8611,399850.9611,796455.9361
+Cholesky Decomposition,9.36492192743764,9.34403475641803,9.56907852713522,9.80909425049519,10.2310240922121,10.5599578268641,20.5268554334554,18.9044681818182,38.0992779120879,24.0629778412377,56.8469825431035,51.9288894736842,75.7380010989011,108.977733333333,115.951276190476,164.127204166667,159.111014285714,308.9641,744.5066,1928.2611,4556.6861,8793.8111,22571.9111
+Matrix Inversion,35.4186902777778,40.902868115942,47.9163,54.8531996732026,148.281957142857,167.517728787879,249.205175,203.305725,349.706475,546.908016666667,1290.6111,1989.2361,1772.6861,3060.2361,3313.0611,3773.2111,5076.7111,8151.3111,14452.4361,32078.7111,75226.5361,148996.5361,412536.3611
+Linear System Solution,14.088697826087,16.7595307330827,17.7408430589152,19.9882126384064,25.2543527811212,27.7271542016807,39.08146,63.2368029411765,79.4382971153846,142.839411111111,395.468716666667,559.54835,586.2566,689.4776,990.87335,1330.83185,1652.6861,2372.9861,3591.3861,8100.1611,19115.3861,34749.7111,92428.4861
+Linear Least Squares,22.6206243632337,22.2334463455149,24.3407333333333,25.8739648648649,32.5984442528736,32.6289902298851,93.8506454545455,107.598372222222,136.922352597403,143.471633333333,254.182808333333,455.648225,638.23435,645.5691,921.87335,1426.5861,1511.8361,2171.1361,4279.6361,10223.9861,27535.5111,53760.6611,129590.1361
+Squared Distance Matrix,23.843823345526,12.5056362305411,12.5595229093567,13.4393796446554,14.4632872960373,15.7837617254793,18.6901065384615,22.4582032920568,61.058930952381,87.2329944444445,138.8275375,149.646508333333,341.430175,328.579391666667,525.388516666667,627.50935,967.04835,1403.5776,2612.6861,6036.5861,14016.4611,25802.9361,56687.7861
+K-Means Run Time,90.7423227272727,85.4807553030303,89.9279181818182,88.6233727272727,89.97345,91.48595,99.1819613636364,102.07865,115.592106944444,139.605608928571,176.2286,228.7489,400.488433333333,559.158016666667,581.157933333333,908.79435,1220.18185,1968.0861,3413.8111,5837.5861,13406.8611,33869.0361,63057.6361
