--- RunTimeJulia.csv
@@ -1,16 +1,0 @@
-912.2244897959183
-164.9055164319249
-615.1685393258427
-781.025641025641
-3077.777777777778
-4428.571428571428
-8966.666666666666
-18401.0
-22000.0
-24499.0
-1359.9
-26700.0
-3962.625
-7433.333333333333
-2511.1111111111113
-269900.0
