--- RunTimeJuliaTable.csv
@@ -1,17 +1,0 @@
-Function Name,Matrix Size,Run Time
-Matrix Generation,10,912.2244897959183
-Matrix Addition,10,164.9055164319249
-Matrix Multiplication,10,615.1685393258427
-Matrix Quadratic Form,10,781.025641025641
-Matrix Reductions,10,3077.777777777778
-Element Wise Operations,10,4428.571428571428
-Matrix Exponential,10,8966.666666666666
-Matrix Square Root,10,18401.0
-Singular Value Decomposition,10,22000.0
-Eigen Decomposition,10,24499.0
-Cholesky Decomposition,10,1359.9
-Matrix Inversion,10,26700.0
-Linear System Solution,10,3962.625
-Linear Least Squares,10,7433.333333333333
-Squared Distance Matrix,10,2511.1111111111113
-K-Means,10,269900.0
