--- vMatrixSizeFull.csv
@@ -1,1 +1,0 @@
-2,5,10,20,50,100,200,300,500,750,1000,2000,3000,4000
